refactor: change UnreachableWalker.cpp to UnreachableWalker.hpp to include hpp file not cpp
